LayoutTests/inspector/geolocation-success.html

<html>
<head>
<script src="../http/tests/inspector/inspector-test.js"></script>
<script>

var mockLatitude = 51.478;
var mockLongitude = -0.166;
var mockAccuracy = 100;

if (window.testRunner) {
    testRunner.setGeolocationPermission(true);
    testRunner.setMockGeolocationPosition(mockLatitude,
                                                    mockLongitude,
                                                    mockAccuracy);
} else
    debug('This test can not be run without the LayoutTestController');

function test()
{
    function callbackComplete()
    {
        InspectorTest.completeTest();
    }
    function callback()
    {
        PageAgent.clearGeolocationData();
        InspectorTest.evaluateInPage("navigator.geolocation.getCurrentPosition(function(pos){console.log('lat: ' + pos.coords.latitude + ',' + ' long: ' + pos.coords.longitude);}, function(err){console.log(err);}, [])", callbackComplete);
    }
    PageAgent.setGeolocationData(43.476093, -80.540299, 150, "None");
    InspectorTest.evaluateInPage("navigator.geolocation.getCurrentPosition(function(pos){console.log('lat: ' + pos.coords.latitude + ',' + ' long: ' + pos.coords.longitude);}, function(err){console.log(err);}, [])", callback);
}


</script>
</head>

<body onload="runTest()">
<p>
</p>

</body>
</html>
