LayoutTests/inspector/geolocation-success-expected.txt

CONSOLE MESSAGE: line 1: lat: 43.476093, long: -80.540299
CONSOLE MESSAGE: line 1: lat: 51.478, long: -0.166
